requirements.txt
Pillow>=10.0.0
ttkbootstrap>=1.10.1
ttkbootstrap-icons>=4.0.0
ttkbootstrap-icons-bs>=1.0.0
pyinstaller>=6.0.0
